fix: give resource loader explicit file ownership
